+++ src/renderer/src/components/devices/DeviceTable.jsx
@@ -1,5 +1,5 @@
 /* eslint-disable no-unused-vars */
-import { Badge, ConfigProvider } from 'antd'
+import { App, Badge, ConfigProvider } from 'antd'
 import { CheckOutlined, CloseOutlined } from '@ant-design/icons'
 import React, { useEffect, useState } from 'react'
 import { useTheme } from 'antd-style'
@@ -8,6 +8,21 @@ import RowContextMenu from '../RowContextMenu/RowContextMenu'
 import { useCallback } from 'react'
 import { useDispatch, useSelector } from 'react-redux'
 import { selectDiscoveryTable, discoverySelector } from '../../features/discoverySlice'
+import { initDeviceAdvanced } from '../../features/deviceAdvanceSettingSlice'
+import { snmpSelector } from '../../features/Preferences/snmpSlice'
+import { initSingleNetworkSetting } from '../../features/singleNetworkSettingSlice'
+import { openAdvanceDrawer } from '../../features/deviceAdvanceSettingSlice'
+import {
+  requestDeviceBeep,
+  requestDeviceReboot,
+  requestOpenTelnet
+} from '../../features/deviceBasiceOperatorSlice'
+import { requestOpenWebData } from '../../features/openWebSlice'
+import { openDialog } from '../../features/dialogSlice'
+import { requestGetBackupRestoreData } from '../../features/singleBackupRestoreSlice'
+import { openDrawer } from '../../features/singleNetworkSettingSlice'
+import { openPortInfoDrawer, initPortInfoData } from '../../features/portInformationSlice'
+import { requestCheckSNMP } from '../../features/discoverySlice'
 // import EnhanceCheckBox from './EnhanceCheckBox/EnhanceCheckBox'
 
 const columns = [
@@ -82,8 +97,18 @@ const DeviceTable = ({ deviceData = [] }) => {
   const [xPos, setXPos] = useState(0)
   const [yPos, setYPos] = useState(0)
   const [showMenu, setShowMenu] = useState(false)
+  const [contextRecord, setContextRecord] = useState({})
   const token = useTheme()
   const dispatch = useDispatch()
+  const { isPrecheck } = useSelector(snmpSelector)
+  const { modal } = App.useApp()
+
+  const showCheckSNMPFailModal = () => {
+    modal.error({
+      title: 'Check SNMP feature fail. ',
+      content: 'Please check SNMP of this device is enable.'
+    })
+  }
   // const [tableType, setTableType] = useState('')
   // const [groupId, setGroupId] = useState()
   // const [order, setOrder] = useState('asc')
@@ -189,6 +214,192 @@ const DeviceTable = ({ deviceData = [] }) => {
       document.addEventListener('click', handleClick)
     }
   })
+
+  const handleItemClick = (key, data) => {
+    console.log(data)
+    const { IPAddress, MACAddress, model, deviceType } = data
+    switch (key) {
+      case 'openOnOSbrowser':
+        window.electron.shell.openExternal(`http://${IPAddress}`)
+        break
+      case 'openOnNMUApplication':
+        return handleOpenWeb(IPAddress, MACAddress)
+      case 'telnet':
+        if (model === 'Cisco CGS2520') {
+          return null
+        }
+        return handleOpenTelnet(IPAddress)
+      case 'beep':
+        if (model === 'Cisco CGS2520') {
+          return null
+        }
+        return handleBeep(IPAddress, MACAddress, deviceType)
+      case 'networkSetting':
+        if (model === 'Cisco CGS2520') {
+          return null
+        }
+        console.log(MACAddress, IPAddress, deviceType, model, deviceType)
+        return handleNetworkSetting(MACAddress, IPAddress, deviceType, model, deviceType)
+      case 'reboot':
+        if (model === 'Cisco CGS2520') {
+          return null
+        }
+        return handleReboot(MACAddress, IPAddress, deviceType)
+      case 'deviceAdvancedSetting':
+        if (model === 'Cisco CGS2520') {
+          return null
+        }
+        return handleDeviceAdvancedSetting(MACAddress, IPAddress, deviceType)
+      case 'portInformation':
+        if (model === 'Cisco CGS2520') {
+          return null
+        }
+        return handlePortInformation(MACAddress, IPAddress, deviceType)
+      case 'backupAndRestore':
+        if (model === 'Cisco CGS2520') {
+          return null
+        }
+        return handleBackupConfig(MACAddress, IPAddress, deviceType)
+      default:
+        break
+    }
+  }
+
+  const handleOpenTelnet = (IPAddress) => {
+    dispatch(requestOpenTelnet(IPAddress))
+  }
+
+  const handleOpenWeb = (IPAddress, MACAddress) => {
+    dispatch(
+      requestOpenWebData({
+        IPAddress,
+        MACAddress
+      })
+    )
+    dispatch(openDialog('webBrowser'))
+  }
+
+  const handleBeep = (IPAddress, MACAddress, deviceType) => {
+    modal
+      .confirm({
+        title: 'Confirm',
+        content: 'This will let device beep.'
+      })
+
+      .then(
+        () => {
+          dispatch(
+            requestDeviceBeep({
+              IPAddress,
+              MACAddress,
+              deviceType
+            })
+          )
+          return null
+        },
+        () => {}
+      )
+      .catch()
+  }
+
+  const handleReboot = async (MACAddress, IPAddress, deviceType) => {
+    const confirm = await modal.confirm({
+      title: 'Confirm',
+      content: 'This will reboot the device.'
+    })
+    console.log(confirm)
+      ? setTimeout(async () => {
+          const confirmed = await modal.success({
+            title: 'Success !',
+            type: 'success',
+            content: 'Device reboot success.'
+          })
+          console.log(confirmed)
+        }, 3000)
+      : setTimeout(async () => {
+          const confirmed = await modal.error({
+            title: 'Error !',
+            type: 'error',
+            content: 'Device reboot fails.'
+          })
+          console.log(confirmed)
+        }, 3000)
+    dispatch(
+      requestDeviceReboot({
+        MACAddress,
+        IPAddress,
+        deviceType
+      })
+    )
+  }
+
+  const handleNetworkSetting = (MACAddress, IPAddress, deviceType) => {
+    dispatch(openDrawer(true), dispatch(openDialog('singleNetworkSetting')))
+    if (deviceType !== 'gwd' || !isPrecheck) {
+      dispatch(initSingleNetworkSetting({ MACAddress }))
+    } else {
+      requestCheckSNMP(
+        {
+          MACAddress,
+          IPAddress
+        },
+        () => {
+          dispatch(initSingleNetworkSetting({ MACAddress }))
+        }
+      )
+    }
+  }
+
+  const handleDeviceAdvancedSetting = (MACAddress, IPAddress, deviceType) => {
+    console.log(MACAddress, IPAddress, deviceType)
+    dispatch(openAdvanceDrawer(true), dispatch(openDialog('advanceSetting')))
+    if (deviceType !== 'gwd' || !isPrecheck) {
+      dispatch(initDeviceAdvanced({ MACAddress }))
+    } else {
+      requestCheckSNMP(
+        {
+          MACAddress,
+          IPAddress
+        },
+        () => {
+          dispatch(initDeviceAdvanced({ MACAddress }))
+        }
+      )
+    }
+  }
+
+  const handlePortInformation = (MACAddress, IPAddress, deviceType) => {
+    dispatch(openPortInfoDrawer(true), dispatch(openDialog('portInformation')))
+    if (deviceType !== 'gwd') {
+      dispatch(initPortInfoData({ MACAddress }))
+    } else {
+      dispatch(
+        requestCheckSNMP({ MACAddress, IPAddress }, (result) => {
+          if (result) {
+            dispatch(initPortInfoData({ MACAddress }))
+          } else {
+            dispatch(showCheckSNMPFailModal())
+          }
+        })
+      )
+    }
+  }
+
+  const handleBackupConfig = (MACAddress, IPAddress, deviceType) => {
+    if (deviceType !== 'gwd') {
+      dispatch(requestGetBackupRestoreData({ MACAddress }))
+    } else {
+      dispatch(
+        requestCheckSNMP({ MACAddress, IPAddress }, (result) => {
+          if (result) {
+            dispatch(requestGetBackupRestoreData({ MACAddress }))
+          } else {
+            dispatch(showCheckSNMPFailModal())
+          }
+        })
+      )
+    }
+  }
   // const handleCheckBoxChange = () => {
   //   dispatch(
   //     selectDiscoveryTable({
@@ -263,16 +474,21 @@ const DeviceTable = ({ deviceData = [] }) => {
             persistenceKey: 'device-table',
             persistenceType: 'localStorage'
           }}
-          onRow={(record, rowIndex) => {
+          onRow={(record) => {
             return {
               onContextMenu: (event) => {
                 console.log(event)
+                setContextRecord(record)
                 handleContextMenu(event)
               }
             }
           }}
         />
-        <RowContextMenu position={{ showMenu, xPos, yPos }} />
+        <RowContextMenu
+          position={{ showMenu, xPos, yPos }}
+          record={contextRecord}
+          onMenuClick={handleItemClick}
+        />
       </ConfigProvider>
     </div>
   )
